requirements.txt
# Core dependencies
playwright==1.41.2
pytest==8.0.0
pytest-playwright==0.4.4
pytest-asyncio==0.23.3

# LLM Integration
openai==1.12.0
anthropic==0.18.1

# CLI & Interface
click==8.1.7
rich==13.7.0
streamlit==1.31.0

# Logging & Monitoring
loguru==0.7.2
structlog==24.1.0

# Git Integration
GitPython==3.1.41

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
jinja2==3.1.3
pydantic==2.6.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

